+++ CHANGELOG.md
@@ -18,7 +18,7 @@ The following emojis are used to highlight certain changes:
 
 ### Changed
 
-- `provider` don't call the reprovide throughput callback function when reprovides are disabled [#871](https://github.com/ipfs/boxo/pull/871) 
+- fix(`provider`): don't reprovide if `reprovideInterval` is set to 0 [#871](https://github.com/ipfs/boxo/pull/871) 
 
 ### Removed